llm_service/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.2
transformers==4.41.0
numpy==1.24.4
torch==2.1.0
python-dotenv==1.0.0
